full dispose pattern to resovle timeout

+++ src/DynamoCoreWpf/Views/HomePage/HomePage.xaml.cs
@@ -57,6 +57,8 @@ namespace Dynamo.UI.Views
 
         internal List<GuidedTourItem> GuidedTourItems;
 
+        private bool _disposed = false;
+
         /// <summary>
         /// A helper tool to let us test flows without relying on side-effects
         /// </summary>
@@ -90,8 +92,7 @@ namespace Dynamo.UI.Views
             DataContextChanged += OnDataContextChanged;
 
         }
-
-
+            
         private void InitializeGuideTourItems()
         {
             GuidedTourItems = new List<GuidedTourItem>
@@ -314,7 +315,10 @@ namespace Dynamo.UI.Views
                 LoadGraphs(recentFiles);
             }
 
-            startPage.DynamoViewModel.RecentFiles.CollectionChanged += RecentFiles_CollectionChanged;
+            if (startPage.DynamoViewModel != null && startPage.DynamoViewModel.RecentFiles != null)
+            {
+                startPage.DynamoViewModel.RecentFiles.CollectionChanged += RecentFiles_CollectionChanged;
+            }
         }
 
         private async void SetLocale()
@@ -586,19 +590,50 @@ namespace Dynamo.UI.Views
         #region Dispose
         public void Dispose()
         {
-            DataContextChanged -= OnDataContextChanged;
-            if (startPage != null)
-            {
-                startPage.DynamoViewModel.PropertyChanged -= DynamoViewModel_PropertyChanged;
-                startPage.DynamoViewModel.RecentFiles.CollectionChanged -= RecentFiles_CollectionChanged;
-            }
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
 
-            this.dynWebView.CoreWebView2.NewWindowRequested -= CoreWebView2_NewWindowRequested;
-
-            if (File.Exists(fontFilePath))
+        protected virtual void Dispose(bool disposing)
+        {
+            if (!_disposed)
             {
-                File.Delete(fontFilePath);
+                if (disposing)
+                {
+                    // Unsubscribe from events
+                    DataContextChanged -= OnDataContextChanged;
+                    if (startPage != null)
+                    {
+                        if (startPage.DynamoViewModel != null)
+                        {
+                            startPage.DynamoViewModel.PropertyChanged -= DynamoViewModel_PropertyChanged;
+                            if (startPage.DynamoViewModel.RecentFiles != null)
+                            {
+                                startPage.DynamoViewModel.RecentFiles.CollectionChanged -= RecentFiles_CollectionChanged;
+                            }
+                        }
+                    }
+
+                    if (this.dynWebView != null && this.dynWebView.CoreWebView2 != null)
+                    {
+                        this.dynWebView.CoreWebView2.NewWindowRequested -= CoreWebView2_NewWindowRequested;
+                    }
+
+                    // Delete font file if it exists
+                    if (File.Exists(fontFilePath))
+                    {
+                        File.Delete(fontFilePath);
+                    }
+                }
+
+                // Free any unmanaged resources here
+                _disposed = true;
             }
+        }
+
+        ~HomePage()
+        {
+            Dispose(false);
         }
         #endregion
